Translate index field type code.
REF BL-392

* Applies translation to type code for index field via helper method
* Refactors out helper methods to helper class.
