+++ app/assets/stylesheets/icons.css.erb
@@ -8,6 +8,7 @@
 }
 
 span.book,
+span.book_book_chapter,
 a.facet_book,
 a.facet_books {
   background: transparent url(<%= asset_path 'icons/book.png' %>) no-repeat top left;
@@ -17,6 +18,7 @@ span.conference_proceeding, a.facet_conference_proceeding, a.facet_conference_pr
   background: transparent url(<%= asset_path 'icons/play-circle.png' %>) no-repeat top left;
   padding-left: 26px; 
 }
+span.journal_article,
 span.journal_periodical,
 span.journal,
 span.newspaper_article,
